Extract TableClassMethods into its own file.
